refactor: reorganized schema utils
